fix: code review and integration tests check
